Cast StatusError for comparison with rc

+++ src/installer/corehost/corehost.cpp
@@ -218,7 +218,7 @@ int exe_start(const int argc, const pal::char_t* argv[])
 
             rc = main_fn_v2(argc, argv, host_path_cstr, dotnet_root_cstr, app_path_cstr);
 
-            if (trace::get_error_writer() != nullptr && rc == StatusCode::FrameworkMissingFailure && !set_error_writer_fn)
+            if (trace::get_error_writer() != nullptr && rc == static_cast<int>(StatusCode::FrameworkMissingFailure) && !set_error_writer_fn)
             {
                 pal::string_t url = get_download_url();
                 trace::error(_X("  _ To run this application, you need to install a newer version of .NET Core."));
